requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
PyMySQL==1.1.1
flask-login==0.6.3
cryptography==42.0.7
WTForms==3.1.2
Flask-Admin==1.6.1
Pillow==10.3.0
numpy==1.26.4
